Update username.ps1
added shared mailbox creation function, with recursive name checks and sanitisation of variable after each loop.

Need to add proper sanitisation of addition and removal to distribution group functions when I have access to a proper IDE.

Need to remember to apply the -WhatIf argument as a check before following through with full function, would be useful incase there's already something with the same name in a distribution group.